requirements.txt
absl-py==2.3.0
anyio==4.9.0
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.42
botocore==1.38.42
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmake==4.0.3
colorama==0.4.6
cryptography==45.0.4
deepface==0.0.93
exceptiongroup==1.3.0
filelock==3.18.0
fire==0.7.0
firebase-admin==6.9.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==1.12
gast==0.4.0
gdown==5.2.0
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
lz4==4.4.4
Markdown==3.8.2
MarkupSafe==3.0.2
msgpack==1.1.1
mtcnn==1.0.0
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
retina-face==0.0.17
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
